# shop-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      status:
        description: 0=禁用,1=启用
        type: integer
      updated_at:
        type: string
    type: object
...
  models.OrderCreateRequest:
    properties:
      address_id:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.OrderProductRequest'
        minItems: 1
        type: array
      remark:
        type: string
    required:
    - address_id
    - products
    type: object
  models.OrderProductRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      use_gold:
        type: boolean
    required:
    - product_id
    - quantity
    type: object
  models.OrderStatusUpdateRequest:
    properties:
      remark:
        type: string
      status:
        type: integer
    required:
    - status
    type: object
...
  models.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 爱影视频APP商城系统和第三方站点管理API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 爱影视频商城API
  version: "1.0"
paths:
...
  /api/v1/shop/categories:
    get:
      consumes:
      - application/json
      description: 获取分类列表，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页记录数，默认20
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取分类列表
      tags:
      - 分类
    post:
      consumes:
      - application/json
      description: 创建新分类
      parameters:
      - description: 分类信息
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: 创建分类
      tags:
      - 分类
  /api/v1/shop/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 删除分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 删除分类
      tags:
      - 分类
    get:
      consumes:
      - application/json
      description: 根据ID获取分类详情
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取分类详情
      tags:
      - 分类
    put:
      consumes:
      - application/json
      description: 更新分类信息
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类信息
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 更新分类
      tags:
      - 分类
  /api/v1/shop/categories/{id}/products:
    get:
      consumes:
      - application/json
      description: 获取指定分类下的商品列表
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页记录数，默认20
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取分类下的商品
      tags:
      - 分类
...
  /api/v1/shop/orders:
    get:
      consumes:
      - application/json
      description: 获取当前用户的订单列表
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页记录数，默认20
        in: query
        name: limit
        type: integer
      - description: 状态，默认全部
        in: query
        name: status
        type: integer
      - description: 开始日期，格式：YYYY-MM-DD
        in: query
        name: start_date
        type: string
      - description: 结束日期，格式：YYYY-MM-DD
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取订单列表
      tags:
      - 订单
    post:
      consumes:
      - application/json
      description: 创建新订单
      parameters:
      - description: 订单信息
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建订单
      tags:
      - 订单
  /api/v1/shop/orders/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取订单详情
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取订单详情
      tags:
      - 订单
  /api/v1/shop/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新订单状态
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.OrderStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新订单状态
      tags:
      - 订单
  /api/v1/shop/orders/pay/{id}:
    post:
      consumes:
      - application/json
      description: 支付订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 支付信息
        in: body
        name: payment
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 支付订单
      tags:
      - 订单
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"